fix: Descompuesta la logica de websocket en hooks
